+++ arch/xtensa/src/esp32s3/esp32s3_tim.c
@@ -425,10 +425,10 @@ static void tim_getcounter(struct esp32s3_tim_dev_s *dev, uint64_t *value)
       /* Read value */
 
       value_32  = getreg32(TIMG_T0HI_REG(priv->gid)); /* High 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
       *value  <<= SHIFT_32;
       value_32  = getreg32(TIMG_T0LO_REG(priv->gid)); /* Low 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
     }
   else
     {
@@ -439,10 +439,10 @@ static void tim_getcounter(struct esp32s3_tim_dev_s *dev, uint64_t *value)
       /* Read value */
 
       value_32  = getreg32(TIMG_T1HI_REG(priv->gid)); /* High 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
       *value  <<= SHIFT_32;
       value_32  = getreg32(TIMG_T1LO_REG(priv->gid)); /* Low 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
     }
 }
 
@@ -545,18 +545,18 @@ static void tim_getalarmvalue(struct esp32s3_tim_dev_s *dev, uint64_t *value)
   if (priv->tid == ESP32S3_TIM_TIMER0)
     {
       value_32  = getreg32(TIMG_T0ALARMHI_REG(priv->gid)); /* High 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
       *value  <<= SHIFT_32;
       value_32  = getreg32(TIMG_T0ALARMLO_REG(priv->gid)); /* Low 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
     }
   else
     {
       value_32  = getreg32(TIMG_T1ALARMHI_REG(priv->gid)); /* High 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
       *value  <<= SHIFT_32;
       value_32  = getreg32(TIMG_T1ALARMLO_REG(priv->gid)); /* Low 32 bits */
-      *value   |= (uint64_t)value_32;
+      *value   |= value_32;
     }
 }
