+++ drivers/timers/mcp794xx.c
@@ -27,6 +27,7 @@
 #include <nuttx/config.h>
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <time.h>
 #include <errno.h>
 #include <nuttx/debug.h>
@@ -498,7 +499,7 @@ int up_rtc_settime(FAR const struct timespec *tp)
       return -EAGAIN;
     }
 
-  rtcinfo("Setting time tp=(%d,%d)\n", (int)tp->tv_sec, (int)tp->tv_nsec);
+  rtcinfo("Setting time tp=(%jd,%ld)\n", (intmax_t)tp->tv_sec, tp->tv_nsec);
 
   /* Get the broken out time */
